* stable-5.6:
  Update Orbit to R20191126223242 for 2019-12
  Prepare 5.6.0-SNAPSHOT builds
  EGit v5.6.0.201912041214-rc1
  Check deletion enablement in key listener
  [history] Show TABs in commit short message as blanks in commit table
  Auto-sharing must not connect root or home directories
  Prepare 5.6.0-SNAPSHOT builds
  EGit v5.6.0.201911271000-m3

+++ org.eclipse.egit.ui/src/org/eclipse/egit/ui/internal/dialogs/AbstractConfigureRemoteDialog.java
@@ -415,7 +415,7 @@ public abstract class AbstractConfigureRemoteDialog
 		specViewer.getTable().addKeyListener(new KeyAdapter() {
 			@Override
 			public void keyPressed(KeyEvent e) {
-				if (e.keyCode == SWT.DEL) {
+				if (e.keyCode == SWT.DEL && deleteRefSpecAction.isEnabled()) {
 					doDeleteRefSpecs();
 				}
 			}
